cleanup ICommitCollection
