+++ src/GitVersionCore/Core/Abstractions/Git/IBranchCollection.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 
 namespace GitVersion
 {
