+++ src/GitVersionCore/Core/Abstractions/Git/ICommitCollection.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace GitVersion
+{
+    public interface ICommitCollection : IEnumerable<ICommit>
+    {
+        ICommitCollection QueryBy(CommitFilter commitFilter);
+    }
+}
