+++ src/GitVersionCore/LibGit2/Commit.cs
@@ -1,8 +1,6 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using GitVersion.Helpers;
-using LibGit2Sharp;
 
 namespace GitVersion
 {
@@ -58,47 +56,4 @@ namespace GitVersion
         public virtual string Message => innerObjectId?.Message;
     }
 
-    public class CommitCollection : IEnumerable<ICommit>
-    {
-        private readonly ICommitLog innerCollection;
-        private CommitCollection(ICommitLog collection) => innerCollection = collection;
-
-        protected CommitCollection()
-        {
-        }
-
-        public virtual IEnumerator<ICommit> GetEnumerator()
-        {
-            foreach (var commit in innerCollection)
-                yield return (Commit)commit;
-        }
-
-        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
-
-        internal static CommitCollection FromCommitLog(ICommitLog b) => b is null ? null : new CommitCollection(b);
-        public virtual CommitCollection QueryBy(CommitFilter commitFilter)
-        {
-            static object GetReacheableFrom(object item)
-            {
-                return item switch
-                {
-                    Commit c => (LibGit2Sharp.Commit)c,
-                    Branch b => (LibGit2Sharp.Branch)b,
-                    _ => null
-                };
-            }
-
-            var includeReachableFrom = GetReacheableFrom(commitFilter.IncludeReachableFrom);
-            var excludeReachableFrom = GetReacheableFrom(commitFilter.ExcludeReachableFrom);
-            var filter = new LibGit2Sharp.CommitFilter
-            {
-                IncludeReachableFrom = includeReachableFrom,
-                ExcludeReachableFrom = excludeReachableFrom,
-                FirstParentOnly = commitFilter.FirstParentOnly,
-                SortBy = (LibGit2Sharp.CommitSortStrategies)commitFilter.SortBy,
-            };
-            var commitLog = ((IQueryableCommitLog)innerCollection).QueryBy(filter);
-            return FromCommitLog(commitLog);
-        }
-    }
 }
