+++ src/GitVersionCore/LibGit2/ObjectId.cs
@@ -1,4 +1,4 @@
-﻿using GitVersion.Helpers;
+using GitVersion.Helpers;
 
 namespace GitVersion
 {
