+++ src/GitVersionCore/LibGit2/ReferenceCollection.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
