+++ src/GitVersionCore/LibGit2/Remote.cs
@@ -1,4 +1,4 @@
-﻿using System.Linq;
+using System.Linq;
 
 namespace GitVersion
 {
