+++ src/GitVersionCore/LibGit2/Tag.cs
@@ -1,4 +1,4 @@
-﻿using LibGit2Sharp;
+using LibGit2Sharp;
 
 namespace GitVersion
 {
